refactor: 예약이 존재하는 예약 시간 삭제 경우 예외 처리

+++ src/main/java/roomescape/time/ReservationTimeService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.exception.ErrorCode;
 import roomescape.exception.RoomescapeException;
+import roomescape.reservation.ReservationRepository;
 import roomescape.time.dto.ReservationTimeRequest;
 import roomescape.time.dto.ReservationTimeResponse;
 
@@ -12,9 +13,11 @@ import roomescape.time.dto.ReservationTimeResponse;
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository, ReservationRepository reservationRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public ReservationTimeResponse create(ReservationTimeRequest reservationTimeRequest) {
@@ -34,8 +37,9 @@ public class ReservationTimeService {
 
     public void delete(Long id) {
         reservationTimeRepository.findById(id).orElseThrow(() -> new RoomescapeException(ErrorCode.RESERVATION_TIME_NOT_FOUND));
+        if (reservationRepository.existsByTimeId(id)) {
+            throw new RoomescapeException(ErrorCode.RESERVATION_TIME_IN_USE);
+        }
         reservationTimeRepository.deleteById(id);
     }
-
-
 }
